relay design refactoring

+++ ConsoleApplication9/IDataPipe.fs
@@ -2,4 +2,4 @@
 
 [<AllowNullLiteral>]
 type IDataPipe= // implemented by objects that receive and send data
-    abstract TotalTransferedData: unit -> int // return the count of the bytes that have been passed through this object
+    abstract TotalTransferedData: unit -> uint64 // return the count of the bytes that have been passed through this object
